+++ backend/canisters/community/impl/src/lifecycle/post_upgrade.rs
@@ -1,7 +1,8 @@
 use crate::jobs::import_groups::finalize_group_import;
 use crate::lifecycle::{init_env, init_state};
 use crate::memory::{get_stable_memory_map_memory, get_upgrades_memory};
-use crate::{Data, read_state};
+use crate::model::events::CommunityEventInternal;
+use crate::{Data, mutate_state, read_state};
 use canister_api_macros::post_upgrade;
 use canister_logger::LogEntry;
 use canister_tracing_macros::trace;
@@ -9,6 +10,7 @@ use community_canister::post_upgrade::Args;
 use instruction_counts_log::InstructionCountFunctionId;
 use stable_memory::get_reader;
 use tracing::info;
+use types::{EventIndex, EventWrapperInternal, MemberJoined};
 
 #[post_upgrade(msgpack = true)]
 #[trace]
@@ -40,5 +42,43 @@ fn post_upgrade(args: Args) {
         state
             .data
             .record_instructions_count(InstructionCountFunctionId::PostUpgrade, now);
+    });
+
+    // TODO: Delete after communities are upgraded
+    mutate_state(|state| {
+        // Read all the existing community events onto the heap
+        let mut community_events = state.data.events.read_all_events();
+
+        // Check that all events have an event index corresponding to their index in the vector
+        for (index, event) in community_events.iter().enumerate() {
+            if event.index != EventIndex::from(index as u32) {
+                panic!("Event index mismatch: expected {}, got {}", index, event.index);
+            }
+        }
+
+        // Append MemberJoined events to CommunityEvents
+        for (user_id, date_joined) in state.data.members.members_joined() {
+            let event = CommunityEventInternal::MemberJoined(Box::new(MemberJoined {
+                user_id,
+                invited_by: None,
+            }));
+            community_events.push(EventWrapperInternal {
+                index: EventIndex::from(community_events.len() as u32),
+                timestamp: date_joined,
+                expires_at: None,
+                event,
+            });
+        }
+
+        // Stable sort the community events by timestamp - events with the same timestamp will stay in the same order
+        community_events.sort_by_key(|event| event.timestamp);
+
+        // Iterate through the community events and rewrite the event indexes to match the vector index
+        for (index, event) in community_events.iter_mut().enumerate() {
+            event.index = EventIndex::from(index as u32);
+        }
+
+        // Write the updated community events back to stable memory
+        state.data.events.overwrite_all_events(community_events);
     });
 }