+++ backend/canisters/community/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Changed
+
+- Introduce and backdate MemberJoined/Left events ([#8280](https://github.com/open-chat-labs/open-chat/pull/8280))
+
 ## [[2.0.1817](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1817-community)] - 2025-07-02
 
 ### Added